mas de roles, avanzado, pero falta la pregunta del mail

+++ beneficiaries.php
@@ -7,7 +7,9 @@
 			<h2>Beneficiarios</h2>
             
             <ul class="toolbar">
-            <?php if(isAnyRol($_SESSION['dms_id'])== 1){?>
+            <?php 
+			$rol=isAnyRol($_SESSION['dms_id']);
+			if($rol== 1){?>
 			<li><a href="includes/forms/sheltersAdd.php" class="btn colorbox">Agregar Beneficiario</a></li>
 			<?php } ?>
                 
@@ -44,10 +46,15 @@
                     </td>
                     <td><?php echo $shelter['email']; ?></td>
                     <td>
+                    <?php 
+					$rol=isAnyRol($_SESSION['dms_id']);
+					if($rol== 1){?>
                     	<ul class="table-actions">
                         	<li><a href="includes/forms/sheltersEdit.php?e=<?php echo $shelter['id']; ?>" class="icon edit colorbox" title="Editar"><span>Editar</span></a></li>
                             <li><a href="includes/forms/delete.php?t=shelters&d=<?php echo $shelter['id']; ?>" class="icon delete colorbox" title="Eliminar"><span>Eliminar</span></a></li>
                         </ul>
+
+					<?php } ?>
                     </td>
                 </tr>
                 <?php
