+++ categories.php
@@ -6,7 +6,9 @@
 <?php if(isset($_POST['bt-delete'])) list($warning, $success) = delete($_POST);?>
 			<h2>Categorias</h2>
 			<ul class="toolbar">
-            <?php if(isAnyRol($_SESSION['dms_id'])== 1){?>
+            <?php 
+			$rol=isAnyRol($_SESSION['dms_id']);
+			if($rol== 1 || $rol== 3 || $rol== 5 || $rol== 6){?>
             	<li><a href="includes/forms/categoriesAdd.php" class="btn colorbox">Agregar Categoria</a></li>
 			<?php } ?>
             </ul>
@@ -28,10 +30,14 @@
                 	<td><?php echo $category['name']; ?></td>
                     <td><?php echo $category['description']; ?></td>
                     <td>
+                     <?php if($rol== 1 || $rol== 3 || $rol== 5 || $rol== 6){?>
                     	<ul class="table-actions">
                         	<li><a href="includes/forms/categoriesEdit.php?e=<?php echo $category['id']; ?>" class="icon edit colorbox" title="Editar"><span>Editar</span></a></li>
                             <li><a href="includes/forms/delete.php?t=categories&d=<?php echo $category['id']; ?>" class="icon delete colorbox" title="Eliminar"><span>Eliminar</span></a></li>
                         </ul>
+					<?php } ?>
+
+        
                     </td>
                 </tr>
                 <?php
